converted multiple system and testing...

+++ src/error.rs
@@ -37,7 +37,7 @@ pub enum ContractError {
     IncorrectFunds {},
 
     #[error("Must send exactly {ticket_price} SEI to enter the raffle")]
-    PayError { ticket_price: u32 },
+    PayError { ticket_price: u64 },
 
     #[error("No participants in the raffle")]
     NoParticipants {},
